# api/docs/swagger.yaml
definitions:
...
  controllers.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      timestamp:
        type: string
    type: object
...
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  controllers.LoginResponse:
    properties:
      container_id:
        type: string
      email:
        type: string
      last_active:
        type: string
      token:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
      wallet_address:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.RegisterResponse:
    properties:
      container_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      token:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  controllers.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  controllers.UpdateUserRequest:
    properties:
      email:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  controllers.UserProfileResponse:
    properties:
      container_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      last_active:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/users/{userID}:
    delete:
      description: Deletes the authenticated user's account and associated container
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Delete user account
      tags:
      - User Management
    get:
      description: Returns user information by ID (only accessible by the user themselves)
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/controllers.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get user by ID
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's profile information
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Updated user data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateUserRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated user information
          schema:
            $ref: '#/definitions/controllers.UserProfileResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Update user information
      tags:
      - User Management
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password, returns a token
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/controllers.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Login user
      tags:
      - Authentication
  /api/v1/users/profile:
    get:
      description: Returns the authenticated user's profile information
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/controllers.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get user profile
      tags:
      - User Management
  /api/v1/users/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and spins up a dedicated Docker container
        with a Jarvis agent instance
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/controllers.RegisterResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Register a new user
      tags:
      - User Management
swagger: "2.0"